SDL_GameObject::getRect() can throw a segfault (unknown):bug:

+++ src/entity/ProjectileHandler.cpp
@@ -44,16 +44,21 @@ bool ProjectileHandler::collision(SDLGameObject *p1)
     if(m_projectiles.empty())
         return false;
 
+    if(p1 == NULL)
+        return false;
+
+    // TODO segfault sometimes with SDLGameObject::getRect
     for(int i = 0; i < m_projectiles.size(); i++)
     {
-        SDLGameObject* p2 = static_cast<SDLGameObject*>(m_projectiles[i]);
+        //SDLGameObject* p2 = static_cast<SDLGameObject*>(m_projectiles[i]);
 
         SDL_Rect r1 = p1->getRect();
-        SDL_Rect r2 = p2->getRect();
+        SDL_Rect r2 = m_projectiles[i]->getRect();
+
         if(SDL_HasIntersection(&r1, &r2))
         {
             p1->kill();
-            p2->kill();
+            m_projectiles[i]->kill();
             return true;
         }
     }
@@ -66,7 +71,7 @@ void ProjectileHandler::checkForCollisions(std::vector<SDLGameObject*> gameObjec
     {
         for(int j = 0; j < m_projectiles.size(); i++)
         {
-            SDLGameObject* pProjectile = dynamic_cast<SDLGameObject*>(m_projectiles[j]);
+            SDLGameObject* pProjectile = static_cast<SDLGameObject*>(m_projectiles[j]);
 
             SDL_Rect r1 = gameObjects[i]->getRect();
             SDL_Rect r2 = pProjectile->getRect();
